#27 - Move video helpers to script file
